@@ -3,7 +3,7 @@ description: A new Flutter project.
 
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
-version: 1.1.0+5
+version: 2.0.0+6
 
 environment:
    sdk: '>=3.0.0 <4.0.0'
@@ -15,38 +15,41 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.8
-  http: ^0.13.5
-  shared_preferences: ^2.2.3
-  flutter_bloc: ^8.1.1
-  firebase_auth: ^4.20.0
-  firebase_core: ^2.32.0
-  cloud_firestore: ^4.17.5
-  equatable: ^2.0.5
-  path_provider: ^2.1.0
+  http: ^1.6.0
+  shared_preferences: ^2.5.3
+  flutter_bloc: ^8.1.6
+  firebase_auth: ^5.7.0
+  firebase_core: ^3.15.2
+  cloud_firestore: ^5.6.12
+  equatable: ^2.0.7
+  path_provider: ^2.1.5
   path: ^1.9.0
-  share_plus: ^7.2.2
+  share_plus: ^10.1.4
   custom_widgets:
     git:
       url: https://github.com/rojanshr1996/custom_widgets.git
       ref: main
-  provider: ^6.1.2
-  firebase_storage: ^11.7.7
-  cached_network_image: ^3.2.3
-  file_picker: ^8.0.6
-  intl: ^0.19.0
+  provider: ^6.1.5
+  firebase_storage: ^12.4.10
+  cached_network_image: ^3.4.1
+  file_picker: ^8.3.7
+  intl: ^0.20.2
   image_picker: ^0.8.9
-  pdf_render: ^1.4.10
-  google_sign_in: ^5.4.4
-  font_awesome_flutter: ^10.7.0
-  like_button: ^2.0.5
-  url_launcher: ^6.2.6
-  flutter_email_sender: ^5.2.0
-  firebase_messaging: ^14.9.4
+  pdfx: ^2.9.2
+  google_sign_in: ^6.3.0
+  font_awesome_flutter: ^10.9.1
+  like_button: ^2.1.0
+  url_launcher: ^6.3.2
+  flutter_email_sender: ^7.0.0
+  firebase_messaging: ^15.2.10
   flutter_screenutil: ^5.9.3
-  firebase_remote_config: ^4.4.7
-  package_info_plus: ^8.0.0
-  get_it: ^7.7.0
-  firebase_crashlytics: ^3.5.7
+  firebase_remote_config: ^5.5.0
+  package_info_plus: ^8.3.1
+  get_it: ^8.3.0
+  firebase_crashlytics: ^4.3.10
+  pdf: ^3.11.1
+  printing: ^5.13.4
+  permission_handler: ^11.3.1
 
 flutter_launcher_icons:
   android: "launcher_icon"
@@ -59,9 +62,9 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_lints: ^2.0.1
-  flutter_launcher_icons: ^0.13.1
-  bloc_test: ^9.1.3
+  flutter_lints: ^3.0.2
+  flutter_launcher_icons: ^0.14.4
+  bloc_test: ^9.1.7
   test: ^1.20.0
 
 flutter:
