@@ -17,7 +17,7 @@ import 'package:notester/widgets/logo_widget.dart';
 import 'package:notester/widgets/safe_area_header.dart';
 
 class SignupScreen extends StatefulWidget {
-  const SignupScreen({Key? key}) : super(key: key);
+  const SignupScreen({super.key});
 
   @override
   _SignupScreenState createState() => _SignupScreenState();
@@ -102,7 +102,10 @@ class _SignupScreenState extends State<SignupScreen> {
                           child: Padding(
                             padding: const EdgeInsets.symmetric(
                                 horizontal: 20, vertical: 2),
-                            child: LogoWidget(height: 50.h),
+                            child: LogoWidget(
+                              height: 50.h,
+                              path: 'assets/notesterIcon.png',
+                            ),
                           ),
                         ),
                       ),
@@ -200,7 +203,7 @@ class _SignupScreenState extends State<SignupScreen> {
                                             value: value!,
                                             fieldName: "Phone",
                                             maxCharacter: 10,
-                                            minCharater: 7),
+                                            minCharacter: 7),
                                         suffixIcon: const Icon(Icons.phone,
                                             color: AppColors.cDarkBlue),
                                       ),
