@@ -1,24 +1,55 @@
-import 'dart:io';
-
+import 'package:flutter/material.dart';
+import 'package:flutter_cache_manager/flutter_cache_manager.dart';
 import 'package:notester/model/model.dart';
 import 'package:notester/widgets/no_data_widget.dart';
 import 'package:notester/widgets/simple_circular_loader.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_cache_manager/flutter_cache_manager.dart';
-import 'package:pdf_render/pdf_render_widgets.dart';
+import 'package:pdfx/pdfx.dart';
 
 class PdfViewerScreen extends StatefulWidget {
   final Args arguments;
-  const PdfViewerScreen({Key? key, required this.arguments}) : super(key: key);
+  const PdfViewerScreen({super.key, required this.arguments});
 
   @override
   _PdfViewerScreenState createState() => _PdfViewerScreenState();
 }
 
 class _PdfViewerScreenState extends State<PdfViewerScreen> {
+  PdfControllerPinch? _pdfController;
+  bool _isLoading = true;
+  bool _hasError = false;
+
   @override
   void initState() {
     super.initState();
+    _loadPdf();
+  }
+
+  Future<void> _loadPdf() async {
+    try {
+      final file =
+          await DefaultCacheManager().getSingleFile(widget.arguments.fileUrl);
+      _pdfController = PdfControllerPinch(
+        document: PdfDocument.openFile(file.path),
+      );
+      if (mounted) {
+        setState(() {
+          _isLoading = false;
+        });
+      }
+    } catch (e) {
+      if (mounted) {
+        setState(() {
+          _isLoading = false;
+          _hasError = true;
+        });
+      }
+    }
+  }
+
+  @override
+  void dispose() {
+    _pdfController?.dispose();
+    super.dispose();
   }
 
   @override
@@ -35,17 +66,18 @@ class _PdfViewerScreenState extends State<PdfViewerScreen> {
           overflow: TextOverflow.ellipsis,
         ),
       ),
-      body: FutureBuilder<File>(
-          future: DefaultCacheManager().getSingleFile(widget.arguments.fileUrl),
-          builder: (context, snapshot) =>
-              snapshot.connectionState == ConnectionState.waiting
-                  ? const Center(child: SimpleCircularLoader())
-                  : snapshot.hasData
-                      ? PdfViewer.openFile(
-                          snapshot.data!.path,
-                        )
-                      : const NoDataWidget(
-                          title: "Error opening PDF. Please try again.")),
+      body: _isLoading
+          ? const Center(child: SimpleCircularLoader())
+          : _hasError
+              ? const NoDataWidget(
+                  title: "Error opening PDF. Please try again.")
+              : PdfViewPinch(
+                  controller: _pdfController!,
+                  padding: 10,
+                  backgroundDecoration: BoxDecoration(
+                    color: Theme.of(context).scaffoldBackgroundColor,
+                  ),
+                ),
     );
   }
 }
